requirements.txt
RPI.GPIO==0.7.1
websockets==12.0
